refactor(eng-analytics): split logic and its tests by domain
